remove gulp, do all file building via webpack
