+++ package.json
@@ -43,15 +43,11 @@
     "eslint-plugin-promise": "4.0.1",
     "eslint-plugin-standard": "4.0.0",
     "extract-text-webpack-plugin": "4.0.0-beta.0",
-    "gulp": "3.9.1",
-    "gulp-load-plugins": "1.5.0",
     "humanize": "0.0.9",
     "jquery": "3.3.1",
-    "js-string-escape": "1.0.1",
     "moment": "2.22.2",
     "node-sass": "4.10.0",
     "pre-commit": "1.2.2",
-    "run-sequence": "2.2.0",
     "sass-loader": "7.0.3",
     "srcdoc-polyfill": "1.0.0",
     "standard": "12.0.1",
