+++ scripts/build.sh
@@ -7,6 +7,5 @@ rm -rf ./dist/
 ec "Successfully removed existing build!" "Failed to remove existing build!"
 
 dps "Building..."
-# ./scripts/build.js
-gulp build
+./scripts/build.js
 ec "Successfully built!" "Failed to build!"
